% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibullfns.R
\name{weib_cdfsim}
\alias{weib_cdfsim}
\title{Inverse CDF simulation for the Weibull distribution}
\usage{
weib_cdfsim(theta, gamma, n, endtime, tau = NA)
}
\arguments{
\item{theta}{Scale parameter \eqn{\theta}}

\item{gamma}{Shape parameter \eqn{\gamma}}

\item{n}{Sample size}

\item{endtime}{Maximum study time, point at which all participants
are censored}

\item{tau}{Change-point(s) \eqn{\tau}}
}
\value{
Dataset with n participants with a survival time
and censoring indicator.
}
\description{
This function simulates data for the Weibull change-point hazard function
by simulating values of the exponential function and substituting them into
the inverse hazard function. This function allows for up to four
change-points. Gamma is held constant.
}
\examples{
nochangepoint <- weib_cdfsim(theta = 0.5, gamma = 2, n = 10,
  endtime = 20)
onechangepoint <- weib_cdfsim(theta = c(0.05, 0.01), gamma = 2,
  n = 10, endtime = 20, tau = 10)
twochangepoints <- weib_cdfsim(theta = c(0.05, 0.01, 0.05),
  gamma = 2, n = 10, endtime = 20, tau = c(8, 12))

}
