% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfns.R
\name{exp_cdfsim}
\alias{exp_cdfsim}
\title{Inverse CDF simulation for the exponential distribution}
\usage{
exp_cdfsim(theta, n, endtime, tau = NA)
}
\arguments{
\item{theta}{Scale parameter \eqn{\theta}}

\item{n}{Sample size}

\item{endtime}{Maximum study time, point at which all participants
are censored}

\item{tau}{Change-point(s) \eqn{\tau}}
}
\value{
Dataset with n participants with a survival time
and censoring indicator.
}
\description{
This function simulates data for the exponential change-point hazard function
by simulating values of the exponential function and substituting them into
the inverse hazard function. This function allows for up to four
change-points.
}
\examples{
nochangepoint <- exp_cdfsim(theta = 0.05, n = 10, endtime = 20)
onechangepoint <- exp_cdfsim(theta = c(0.05, 0.01), n = 10,
  endtime = 20, tau = 10)
twochangepoints <- exp_cdfsim(theta = c(0.05, 0.01, 0.05),
  n = 10, endtime = 20, tau = c(8, 12))

}
