% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfns.R
\name{exp_icdf}
\alias{exp_icdf}
\title{Inverse CDF for the exponential distribution}
\usage{
exp_icdf(u, theta)
}
\arguments{
\item{u}{Numerical value(s) to be converted to exponential variable(s)}

\item{theta}{Scale parameter \eqn{\theta}}
}
\value{
If inputs are numeric, output is a value or vector of values
from the inverse CDF of the exponential function.
}
\description{
\code{exp_icdf} returns a value from the inverse CDF of the
exponential function.
}
\details{
This function uses the exponential function in the form
\deqn{f(t)=\theta e^{-\theta t}}
to get the inverse CDF
\deqn{F^{-1}(u)=(-log(1-u))/\theta.} It can be
implemented directly and is also called by the functions
\code{\link{exp_memsim}} and \code{\link{exp_cdfsim}}.
}
\examples{
simdta <- exp_icdf(u = runif(500), theta = 0.05)

}
