% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfns.R
\name{exp_memsim}
\alias{exp_memsim}
\title{Memoryless simulation for the exponential distribution}
\usage{
exp_memsim(theta, n, endtime, tau = NA)
}
\arguments{
\item{theta}{Scale parameter \eqn{\theta}}

\item{n}{Sample size}

\item{endtime}{Maximum study time, point at which all participants
are censored}

\item{tau}{Change-point(s) \eqn{\tau}}
}
\value{
Dataset with n participants with a survival time
and censoring indicator.
}
\description{
\code{exp_memsim} returns a dataset simulated from the exponential
hazard distribution with K change-points.
}
\details{
This function simulates data between change-points from independent
exponential distributions using the inverse exponential CDF implemented in
the function \code{exp_icdf}.
}
\examples{
nochangepoint <- exp_memsim(theta = 0.05, n = 10, endtime = 20)
onechangepoint <- exp_memsim(theta = c(0.05, 0.01), n = 10,
  endtime = 20, tau = 10)
twochangepoints <- exp_memsim(theta = c(0.05, 0.01, 0.05),
  n = 10, endtime = 20, tau = c(8, 12))

}
